requirements.txt
-e .

aws-cdk.core
aws-cdk.aws_s3
aws-cdk.aws_ec2
aws-cdk.aws_autoscaling
aws-cdk.aws_elasticloadbalancingv2
aws-cdk.aws_secretsmanager
aws-cdk.aws_rds
aws-cdk.aws_sns_subscriptions
aws-cdk.aws_sqs
aws_cdk.aws_lambda
aws_cdk.aws_logs
aws_cdk.aws_events_targets
aws_cdk.aws_dynamodb
aws_cdk.aws_cloudwatch_actions
# new 0.0.1
